Aurum: Gets meaningful log from lib and fix build error on 64bit system

+++ libaurum/src/Comparer.cc
@@ -45,7 +45,7 @@ std::vector<std::shared_ptr<AccessibleNode>> Comparer::findObjects(const std::sh
 {
     std::list<std::shared_ptr<PartialMatch>> partialList{};
     std::vector<std::shared_ptr<AccessibleNode>> ret = findObjects(root, 0, 0, partialList);
-    LOGI("%d object(s) found", ret.size());
+    LOGI("%d object(s) found", (int)ret.size());
     return ret;
 }
